+++ tht-admin/src/main/java/com/tht/thtadmin/ui/login/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.tht.thtadmin.ui.login;
+package com.tht.thtadmin.ui.login.request;
 
 import jakarta.validation.constraints.NotEmpty;
